Change to use handlers object
